fix(textfield): move the BTF2 overload to its own file so detekt passes
`TextField.kt` hit the 15-function ceiling the moment the BTF2 overload landed in it,
and the rule is right: the file already carried three public overloads, two selector
variants and their four core implementations. The state overload and its CoreTextField
now live in `TextFieldStateField.kt`, which takes the file back to 13.

`TextFieldData` and its `TextFieldSize?.data` extension go from private to internal so
the moved code can still reach them. Neither is visible outside the module, so nothing
changes for a consumer.

The API dump moves the new function from `TextFieldKt` to `TextFieldStateFieldKt`. That
function is new on this branch and has never been released, so no published binary names
either facade, and the dump is still additions-only against main.
